requirements.txt
absl-py==0.7.1
astor==0.8.0
certifi==2019.3.9
chardet==3.0.4
cloudpickle==1.1.1
cycler==0.10.0
decorator==4.4.0
future==0.17.1
gast==0.2.2
gin-config==0.1.3
google-pasta==0.1.6
grpcio==1.20.1
gym==0.12.1
gym-super-mario-bros==7.2.1
h5py==2.9.0
idna==2.8
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
Markdown==3.1
matplotlib==3.1.0
nes-py==6.2.2
numpy==1.16.3
opencv-python==4.1.0.25
Pillow==6.0.0
protobuf==3.7.1
pygame==1.9.6
pyglet==1.3.2
pyparsing==2.4.0
python-dateutil==2.8.0
requests==2.22.0
scipy==1.3.0
six==1.12.0
tb-nightly==1.14.0a20190520
tensorflow-estimator-2.0-preview==1.14.0.dev2019052000
termcolor==1.1.0
tf-agents-nightly==0.2.0.dev20190430
tf-estimator-nightly==1.14.0.dev2019030115
tf-nightly-2.0-preview==2.0.0.dev20190518
tfp-nightly==0.7.0.dev20190520
tqdm==4.32.1
urllib3==1.25.2
Werkzeug==0.15.4
wrapt==1.11.1
